refactor(main): split main.ts 1304→300 LOC into 6 main-commands modules
v1.25.1 Phase C-PR3 — continue the big-file split from PR1 (wiki-engine)
and PR2 (settings). main.ts was 1304 LOC and the next-largest target after
those two. PR3 reduces it by 77% with zero behavior change.

Design choice: mixin pattern (user-confirmed 2026-07-19).

  - 6 mixin modules under src/main-commands/ totaling 883 LOC:
    pdf-cache, connection, schema, query-lint, ingest (5 mixins) +
    command-registry (1 free function, breaks main↔connection circular dep).
  - 1 supporting module in src/core/create-plugin-llm-client.ts (31 LOC)
    extracted from main.ts to break the main↔connection circular dep.
  - main.ts retains: settings lifecycle (loadSettings/saveSettings),
    progress helpers, ingest dispatch, manager construction, and onload.

Mixin pattern: `Object.assign(LLMWikiPlugin.prototype, mixinObj)` +
`export interface LLMWikiPlugin extends MixinMethods` interface merging.
Preserves the `plugin.method()` call surface used by 10 existing tests.

Bug fixed during code-review: `testLLMConnection` extraction initially
dropped the `isWikiInitialized + ensureWikiStructure` block — restored
so first-time users still get wiki folder auto-creation on connect.

Fix: lint Smart Fix All completion Notice UX

User reported: "after Fix All runs, no Notice seems to appear — silent."
Root cause: phaseCount = allResults.length. allResults only pushes when
fixed > 0, so if every phase failed silently, phaseCount === 0, and the
Notice read "All fixes complete. See log for details.: 0 phases."
which users read as "nothing happened."

Fix: branch summary on phasesModified === 0 with distinct
"No changes were made" copy; switch duration NOTICE_ERROR(8s) →
NOTICE_NORMAL(5s) to match other fix callbacks (non-blocking).

i18n: 10 locales × 2 new keys (lintFixAllNoChanges, lintFixPhasesLabel).

Verification

- Gate 1: lint 0/0, tsc 0/0, test 2248/2248 (+12 new), build clean,
  css-lint 0/0.
- Mixin invocation surface: 10 testLLMConnection tests pass unchanged.
- Independent audit (diff-traced, no trust in prior summary):
  * 11 addCommand IDs unchanged (diff = empty).
  * onload 22-step order preserved (registerWikiCommands is a pure
    bundling of the 11 commands + 2 ribbons + 3 callbacks + settingTab
    that originally lived inline; order-sensitive registerView stays
    in onload BEFORE the bundle call).
  * All public method signatures preserved.
  * 4 private fields (ingestQueue/progressNotice/ingestStatusBar/
    batchProgress) promoted to public so mixins can read/write them.
